requirements.txt
mcp>=1.0.0
pyyaml>=6.0
aiohttp>=3.9.0
